terminacion del boton delete, tabla documentsTypes

+++ app/Http/Controllers/DocumentTypeController.php
@@ -103,15 +103,42 @@ class DocumentTypeController extends Controller
         return response()->json($mensaje);
     }
 
-    public function destroy(DocumentType $documentType)
-    {
+    public function destroy(Request $request, $id)
+{
+    $documentType = DocumentType::find($id);
+
+    if ($documentType) {
         $documentType->delete();
+        if ($request->ajax()) {
+            return response()->json([
+                'title' => '¡Eliminación exitosa!',
+                'html' => 'El tipo de documento ha sido eliminado exitosamente',
+                'icon' => 'success'
+            ]);
+        } else {
+            return response()->json([
+                'success' => true,
+                'message' => 'El tipo de documento ha sido eliminado exitosamente',
+                'html' => view('documentType.table')->render(),
+            ]);
+            
+        }
+    } else {
+        if ($request->ajax()) {
+            return response()->json([
+                'title' => '¡Error!',
+                'html' => 'El tipo de documento no existe',
+                'icon' => 'error'
+            ]);
+        } else {
+            return redirect()->route('documentType.index')->with('error', 'El tipo de documento no existe');
+        }
+    }
+}
 
-        Session::flash('message', "usted a eliminado un usuario");
-        return back();
-    } 
 
-    public function edit($id){
+    public function edit($id)
+    {
         
         
         $pepe = DocumentType::find($id);
